rearrange files; rename sublabel animation constants
