+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Vertical Main Screen/RowButton+PaymentMethodMessagingSublabelView.swift
@@ -1,129 +1,11 @@
 //
-//  RowButton+Sublabel.swift
+//  RowButton+PaymentMethodMessagingSublabelView.swift
 //  StripePaymentSheet
 //
 
 @_spi(STP) import StripeCore
 @_spi(STP) import StripeUICore
 import UIKit
-
-// MARK: - SublabelView protocol
-
-extension RowButton {
-
-    fileprivate static let sublabelVisibilityAnimationDuration: TimeInterval = 0.2
-    fileprivate static let sublabelFadeAnimationDuration: TimeInterval = 0.1
-
-    /// Defines the interface for sublabel views displayed beneath the primary label in a `RowButton`.
-    /// Conforming types manage their own visibility transitions and text state.
-    protocol SublabelView: UIView {
-        /// Whether this sublabel variant needs to expand beyond the standard row height.
-        var needsUnlimitedHeight: Bool { get }
-        /// Whether the sublabel currently contains displayable content.
-        var hasText: Bool { get }
-        /// Updates the displayed text, optionally animating the visibility transition.
-        func setSublabel(text: String?, animated: Bool)
-        /// Notifies the sublabel that the parent row's selection state changed.
-        func updateSelectedState(_ isRowSelected: Bool, willDisplayForm: Bool)
-    }
-}
-
-// MARK: - PlainSublabelView
-
-extension RowButton {
-
-    /// A single-line text sublabel used for static descriptive text beneath a payment method name.
-    final class PlainSublabelView: UIView, SublabelView {
-
-        var needsUnlimitedHeight: Bool { false }
-
-        var hasText: Bool {
-            guard let text = textLabel.text else {
-                return false
-            }
-            return !text.isEmpty
-        }
-
-        let textLabel: UILabel
-
-        init(
-            text: String?,
-            appearance: PaymentSheet.Appearance,
-            isEmbedded: Bool
-        ) {
-            self.textLabel = UILabel()
-
-            super.init(frame: .zero)
-
-            if isEmbedded, let customFont = appearance.embeddedPaymentElement.row.subtitleFont {
-                textLabel.font = customFont
-            } else {
-                textLabel.font = appearance.scaledFont(
-                    for: appearance.font.base.regular,
-                    style: .caption1,
-                    maximumPointSize: 20
-                )
-            }
-            textLabel.numberOfLines = 1
-            textLabel.adjustsFontSizeToFitWidth = true
-            textLabel.adjustsFontForContentSizeCategory = true
-            textLabel.text = text
-
-            let textColor: UIColor = {
-                guard isEmbedded else {
-                    return appearance.colors.componentPlaceholderText
-                }
-                switch appearance.embeddedPaymentElement.row.style {
-                case .flatWithRadio, .flatWithCheckmark, .flatWithDisclosure:
-                    return appearance.colors.textSecondary
-                case .floatingButton:
-                    return appearance.colors.componentPlaceholderText
-                }
-            }()
-
-            textLabel.textColor = textColor
-            addAndPinSubview(textLabel)
-            self.isHidden = text?.isEmpty ?? true
-        }
-
-        required init?(coder: NSCoder) {
-            fatalError("init(coder:) has not been implemented")
-        }
-
-        func setSublabel(text: String?, animated: Bool) {
-            guard text != textLabel.text else { return }
-
-            let showDuration = animated ? sublabelVisibilityAnimationDuration : 0
-            let fadeDuration = animated ? sublabelFadeAnimationDuration : 0
-
-            if let text {
-                textLabel.text = text
-                alpha = 0
-                UIView.animate(withDuration: showDuration) {
-                    self.isHidden = text.isEmpty
-                }
-                UIView.animate(
-                    withDuration: fadeDuration,
-                    delay: max(0, showDuration - fadeDuration)
-                ) {
-                    self.alpha = 1
-                }
-            } else {
-                UIView.animate(withDuration: showDuration) {
-                    self.textLabel.text = nil
-                    self.isHidden = true
-                    self.superview?.setNeedsLayout()
-                    self.superview?.layoutIfNeeded()
-                }
-            }
-        }
-
-        func updateSelectedState(_ isRowSelected: Bool, willDisplayForm: Bool) {
-            // Plain sublabel has no selection-dependent behavior
-        }
-
-    }
-}
 
 // MARK: - PaymentMethodMessagingSublabelView
 
@@ -210,26 +92,26 @@ extension RowButton {
         private func expand() {
             promotionTextView.alpha = 0
 
-            UIView.animate(withDuration: sublabelVisibilityAnimationDuration) { [self] in
+            UIView.animate(withDuration: sublabelIsHiddenAnimationDuration) { [self] in
                 self.isHidden = false
                 promotionTextView.isHidden = false
             }
 
             UIView.animate(
-                withDuration: sublabelFadeAnimationDuration,
-                delay: sublabelVisibilityAnimationDuration - sublabelFadeAnimationDuration
+                withDuration: sublabelAlphaAnimationDuration,
+                delay: sublabelIsHiddenAnimationDuration - sublabelAlphaAnimationDuration
             ) { [self] in
                 promotionTextView.alpha = 1
             }
         }
 
         private func collapse() {
-            UIView.animate(withDuration: sublabelVisibilityAnimationDuration) { [self] in
+            UIView.animate(withDuration: sublabelIsHiddenAnimationDuration) { [self] in
                 self.isHidden = true
                 promotionTextView.isHidden = true
             }
 
-            UIView.animate(withDuration: sublabelVisibilityAnimationDuration) { [self] in
+            UIView.animate(withDuration: sublabelIsHiddenAnimationDuration) { [self] in
                 promotionTextView.alpha = 0
             }
         }
