+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Vertical Main Screen/RowButton.swift
@@ -666,3 +666,24 @@ private extension STPPaymentMethod {
         return nil
     }
 }
+
+// MARK: - SublabelView protocol
+
+extension RowButton {
+
+    static let sublabelIsHiddenAnimationDuration: TimeInterval = 0.2
+    static let sublabelAlphaAnimationDuration: TimeInterval = 0.1
+
+    /// Defines the interface for sublabel views displayed beneath the primary label in a `RowButton`.
+    /// Conforming types manage their own visibility transitions and text state.
+    protocol SublabelView: UIView {
+        /// Whether this sublabel variant needs to expand beyond the standard row height.
+        var needsUnlimitedHeight: Bool { get }
+        /// Whether the sublabel currently contains displayable content.
+        var hasText: Bool { get }
+        /// Updates the displayed text, optionally animating the visibility transition.
+        func setSublabel(text: String?, animated: Bool)
+        /// Notifies the sublabel that the parent row's selection state changed.
+        func updateSelectedState(_ isRowSelected: Bool, willDisplayForm: Bool)
+    }
+}
